Move file Editor to CodeEditor
