+++ src/pages/DocPage.tsx
@@ -4,7 +4,7 @@ import { createStyles, makeStyles } from '@material-ui/core/styles';
 
 import NavBar from 'components/NavBar';
 import Toolbar from 'components/Toolbar';
-import Editor from 'components/Editor';
+import Editor from 'components/Editor/CodeEditor';
 
 type DocPageProps = {
   docKey: string;
